Move send_email module to modules directory.
